requirements.txt
Faker==19.11.0
nicegui==1.3.18
python-decouple==3.8
SQLAlchemy==2.0.22
llama-index==0.8.47
langchain==0.0.319
openai==0.28.1
sentence-transformers==2.2.2